Added comments, added bool to platforms, added inspector key choice to buttons, made button prefab
Commented code | created a button prefab ||||    Added a boolean to the platforms so they can be turned on and off by a button as tim asked. If the button space in the inspector is left blank then the platform will always be on |||| Added the keycode option in the button inspector so we can choose what key is used to trigger button. ||| added the platform code to the playermove script. ||| added a new global layer called BehindWall so sprites can be specifically rendered behind the walls. used this in the buttons